// frontend/school-management/src/app/features/auth/models/register.model.ts
import { BloodType } from "../../../core/models/types/blood-type.type";
import { Gender } from "../../../core/models/types/gender.type";

export interface RegisterUser {
  name: string;
  last_name: string;
  email: string;
  password: string;
  phone_number: string;
  birthdate: string;
  gender: Gender | null;
  curp: string;
  address: string[] | null;
  blood_type: BloodType | null;
  status: string;

}